Move pandas and pyarrow version checks
