+++ lib/tests/streamlit/elements/data_editor_test.py
@@ -29,7 +29,11 @@ import pyarrow as pa
 from parameterized import parameterized
 
 import streamlit as st
-from streamlit.dataframe_util import DataFormat, convert_arrow_bytes_to_pandas_df
+from streamlit.dataframe_util import (
+    DataFormat,
+    convert_arrow_bytes_to_pandas_df,
+    is_pandas_version_less_than,
+)
 from streamlit.elements.lib.column_config_utils import (
     INDEX_IDENTIFIER,
     ColumnDataKind,
@@ -46,7 +50,6 @@ from streamlit.elements.widgets.data_editor import (
 )
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.Arrow_pb2 import Arrow as ArrowProto
-from streamlit.type_util import is_pandas_version_less_than
 from tests.delta_generator_test_case import DeltaGeneratorTestCase
 from tests.streamlit.data_mocks import SHARED_TEST_CASES, CaseMetadata
